+++ spec/rubocop/cop/style/string_literals_spec.rb
@@ -296,15 +296,16 @@ describe RuboCop::Cop::Style::StringLiterals, :config do
         }
       end
 
-      it 'does not crash on strings with line breaks in them' do
+      it 'registers an offense for strings with line breaks in them' do
         inspect_source(cop,
                        ['"--',
                         'SELECT *',
                         'LEFT JOIN X on Y',
                         'FROM Models"'])
-        # TODO: We should actually get an offense report here telling us to use
-        # single quotes. For now, we only check that we don't crash.
-        expect(cop.offenses).to be_empty
+        expect(cop.offenses.size).to eq(1)
+        expect(cop.messages).to eq(['Prefer single-quoted strings when you ' \
+                                    "don't need string interpolation or " \
+                                    'special symbols.'])
       end
 
       it 'accepts continued strings using all single quotes' do
@@ -339,6 +340,18 @@ describe RuboCop::Cop::Style::StringLiterals, :config do
 
       it "doesn't register offense for double quotes with embedded single" do
         inspect_source(cop, ['"abc\'" \\',
+                             '"def"'])
+        expect(cop.offenses).to be_empty
+      end
+
+      it 'accepts for double quotes with an escaped special character' do
+        inspect_source(cop, ['"abc\\t" \\',
+                             '"def"'])
+        expect(cop.offenses).to be_empty
+      end
+
+      it 'accepts for double quotes with an escaped normal character' do
+        inspect_source(cop, ['"abc\\!" \\',
                              '"def"'])
         expect(cop.offenses).to be_empty
       end